requirements.txt
# Cloud Orchestration
azure-ai-ml>=1.12.0
azure-identity>=1.15.0

# Core Data Science
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.3.0
joblib>=1.3.0
mlflow>=2.8.0

# System Resources
psutil>=5.9.0
